// src/app/models/asesor.ts
export interface Asesor {
   id: number;
   username: string;
   password: string;
   primerNombre: string;
   segundoNombre: string;
   primerApellido: string;
   segundoApellido: string;
   tipoIdentificacion: string;
   identificacion: string;
   email: string;
   estado: boolean;
   role : string;
}